RI-8200 Split provider constants/utils/types into separate files
Follows the ComponentName.{constants,utils,types}.ts convention:
- Renames commandIds.ts → ProductionWriteConfirmationProvider.constants.ts
  (holds BrowserConfirmationCommandId enum and REDIS_CONFIRMATION_COMMAND_PREFIX).
- Adds ProductionWriteConfirmationProvider.utils.ts with
  toRedisConfirmationCommandId and toCommandIdArray (the latter was inline
  in the .tsx).
- Moves the provider Props interface into .types.ts and renames it to
  ProductionWriteConfirmationProviderProps.
