+++ redisinsight/ui/src/components/production-write-confirmation/ProductionWriteConfirmationProvider.types.ts
@@ -24,3 +24,7 @@ export interface ProductionWriteConfirmationContextValue {
     confirmation: ProductionWriteConfirmationRequest,
   ) => void
 }
+
+export interface ProductionWriteConfirmationProviderProps {
+  children: React.ReactNode
+}
