+++ redisinsight/ui/src/components/production-write-confirmation/index.ts
@@ -4,9 +4,8 @@ export {
 } from './ProductionWriteConfirmationProvider'
 export type {
   ProductionWriteConfirmationContextValue,
+  ProductionWriteConfirmationProviderProps,
   ProductionWriteConfirmationRequest,
 } from './ProductionWriteConfirmationProvider.types'
-export {
-  BrowserConfirmationCommandId,
-  toRedisConfirmationCommandId,
-} from './commandIds'
+export { BrowserConfirmationCommandId } from './ProductionWriteConfirmationProvider.constants'
+export { toRedisConfirmationCommandId } from './ProductionWriteConfirmationProvider.utils'
